requirements.txt
aiohttp==3.7.4
dateparser==0.7.4
dblpy==0.4.0
discord.py==2.0.1
motor==2.3.0
pytz==2020.1
regex==2020.10.23
requests==2.24.0
Unidecode==1.1.1
websockets==8.1
python-dotenv==0.20.0
multidict==4.5.0
pymongo==3.11.0